% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itses.R
\name{itses}
\alias{itses}
\title{Iterative Sparse Estimation}
\usage{
itses(
  y,
  method = "ST",
  m = 5,
  init.lambda = "median",
  max.length = 5e+06,
  minimizationmethod = "numeric",
  debug = FALSE,
  sd = NA,
  sparsity = NA,
  sparse.mad = TRUE,
  remove.zero = TRUE,
  tol = 1e-08,
  h = 0.4,
  ...
)
}
\arguments{
\item{y}{A \code{n} vector of observations.}

\item{method}{Estimator which will be used. Can be either
"ST" (Soft-Threshold estimator) or "HT" (Hard-Threshold estimator).
Defaults to "ST".}

\item{m}{The number of iterative steps. Default is \code{5}.}

\item{init.lambda}{Initial threshold to start iterations at. Can be:
\code{numeric}; "median"; "visu" (universal threshold); and "half-visu"
(universal threshold divded by 2). Defaults to "median".}

\item{max.length}{The maximum number of observations \code{n}. If\code{n} is above
internal calculations will downsample \code{y} to be of \code{max.length} size.
Defaults to \code{5e6}.}

\item{minimizationmethod}{The risk minimization method which will be used.
Can be either "numeric" or "sampling". Default is "numeric".}

\item{debug}{\code{logical}. Specify wether or not to print debug code from
iterations.}

\item{sd}{Standard deviation of \code{y} i.e. noise levels. If \code{NA} will estiamt
e using specified MAD estimator. Default is \code{NA}.}

\item{sparsity}{Sparsity of \code{y} if known. \verb{0 < sparsity < 1}, closer to 1
means highly sparse. Is used in noise estimation if sparse noise estimation
is used. Default is \code{NA}.}

\item{sparse.mad}{\code{logical}. If \code{TRUE} use sparsity in MAD noise estimation.
Default is \code{TRUE}.}

\item{remove.zero}{\code{logical}. Remove observation that are zero from noise
estimatiom. Default is \code{TRUE}.}

\item{tol}{A numeric that determines the sensitivity of threshold selection.
Default is \code{1e-8}, iteration will stop if change is less than set \code{tol}.}

\item{h}{Parameter for deciding eligibility of sparsity measure.
Default is \code{0.4}, only used with SparseMAD estimator.}

\item{...}{parameters passed to risk minmization methods.}

\item{b}{number of samples to take. Default is \code{10}.}

\item{k}{number of thresholds to initially evaluate. Default is \code{10}.}

\item{max_num_iters}{maximum number of iterations with Newton's method or
spline interpolations (depending on method used). Default is \code{10}.}
}
\value{
List object with iteration results.
}
\description{
Performs iterative sparse estimation of many normal means either using
the soft threshold estimator or the hard-threshold estimator.
}
\details{
Given \eqn{Y = (Y1, Y2, \dots, Yn)}, such that:
\deqn{Yi ~ N(\theta i, \sigma2), 1 \le i \le n.}
\eqn{N(\theta, \sigma2)} denotes the normal distribution with
means \eqn{\theta} and variance \eqn{\sigma2}. The means are to
be estimated. Variance is either known or to be estimated.

Means are either estimated using the soft-threhsold (ST) or the
hard-threshold (HT) estimator. ITSES minimize expected square error loss
under repeated sampling (risk).

Soft-threshold: \eqn{\hat{\theta}(x)= sign(x)(abs(x)-\lambda)(abs(x)>\lambda)}.

Hard-threshold: \eqn{\hat{\theta}(x)= x(abs(x)>\lambda)}.
}
