% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise-estimation.R
\name{sparse.mad.estimator}
\alias{sparse.mad.estimator}
\title{SparseMAD noise estimator.}
\usage{
sparse.mad.estimator(y, h = 0.4, debug = FALSE)
}
\arguments{
\item{y}{Vector of length \code{n} estiamte noise from.}

\item{h}{Numeric to decide eligibility of sparsity estimate. Defaults to \code{0.4}.}

\item{debug}{\code{logical}, print logs if \code{TRUE}. Defaults to \code{FALSE}.}
}
\description{
Estiamtes noise using a filtered MAD estimator.
}
